LayoutTests/http/tests/security/xssAuditor/cookie-injection-expected.txt

CONSOLE MESSAGE: line 4: The Set-Cookie meta tag is obsolete and was ignored. Use the HTTP header Set-Cookie or document.cookie instead.
ALERT: PASS
